requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-chroma
chromadb
sentence-transformers
pypdf
groq
python-dotenv